requirements.txt
discord.py>=1.7.2,<2.0.0
GitPython>=3.1.14,<4.0.0
redis>=3.5.3,<4.0.0
